+++ src/components/Navbar.tsx
@@ -1,4 +1,4 @@
-import './Navbar.css';
+import '../styles/Navbar.css';
 
 const Navbar = () => {
     return (
